third_party/WebKit/LayoutTests/fast/encoding/api/utf-round-trip.html

<!DOCTYPE html>
<title>Encoding API: UTF encoding round trips</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/encodings.js"></script>
<script>

var BATCH_SIZE = 0x1000; // Convert in batches spanning this many code points.
var SKIP_SIZE = 0x77; // For efficiency, don't test every code point.

function fromCodePoint(cp) {
    if (0xD800 <= cp && cp <= 0xDFFF) throw new Error('Invalid code point');

    if (cp <= 0xFFFF)
        return String.fromCharCode(cp);

    // outside BMP - encode as surrogate pair
    return String.fromCharCode(0xD800 + ((cp >> 10) & 0x3FF), 0xDC00 + (cp & 0x3FF));
}

function makeBatch(cp) {
    var string = '';
    for (var i = cp; i < cp + BATCH_SIZE && cp < 0x10FFFF; i += SKIP_SIZE) {
        if (0xD800 <= i && i <= 0xDFFF) {
            // surrogate half
            continue;
        }
        string += fromCodePoint(i);
    }
    return string;
 }

utf_encodings.forEach(function(encoding) {
    test(function() {
        for (var i = 0; i < 0x10FFFF; i += BATCH_SIZE) {
            var string = makeBatch(i);
            var encoded = new TextEncoder(encoding).encode(string);
            var decoded = new TextDecoder(encoding).decode(encoded);
            assert_equals(decoded, string);
        }
    }, encoding + ' - encode/decode round trip');
});


// Inspired by:
// http://ecmanaut.blogspot.com/2006/07/encoding-decoding-utf8-in-javascript.html
function encode_utf8(string) {
    var utf8 = unescape(encodeURIComponent(string));
    var octets = [];
    for (var i = 0; i < utf8.length; i += 1)
        octets.push(utf8.charCodeAt(i));
    return octets;
}

function decode_utf8(octets) {
    var utf8 = String.fromCharCode.apply(null, octets);
    return decodeURIComponent(escape(utf8));
}

test(function() {
    for (var i = 0; i < 0x10FFFF; i += BATCH_SIZE) {
        var string = makeBatch(i);
        var expected = encode_utf8(string);
        var actual = new TextEncoder('UTF-8').encode(string);
        assert_array_equals(actual, expected);
    }
}, 'UTF-8 encoding (compare against unescape/encodeURIComponent)');

test(function() {
    for (var i = 0; i < 0x10FFFF; i += BATCH_SIZE) {
        var string = makeBatch(i);
        var encoded = encode_utf8(string);
        var expected = decode_utf8(encoded);
        var actual = new TextDecoder('UTF-8').decode(new Uint8Array(encoded));
        assert_equals(actual, expected);
    }
}, 'UTF-8 decoding (compare against decodeURIComponent/escape)');

</script>
